rename Input to TextInput and use daisy
